Revert "Merge branch 'development' into Camera-Overhaul-Simple"
This reverts commit 94b312852464da5194e4b3367c4377bb1829a0b2, reversing
changes made to bb2c1d8c9f73d3fac1889ccbdc539e437215f6f4.

+++ arcade/gui/events.py
@@ -21,8 +21,8 @@ class UIMouseEvent(UIEvent):
     Covers all mouse event
     """
 
-    x: int
-    y: int
+    x: float
+    y: float
 
     @property
     def pos(self):
@@ -32,8 +32,8 @@ class UIMouseEvent(UIEvent):
 @dataclass
 class UIMouseMovementEvent(UIMouseEvent):
     """Triggered when the mouse is moved."""
-    dx: int
-    dy: int
+    dx: float
+    dy: float
 
 
 @dataclass
@@ -46,8 +46,8 @@ class UIMousePressEvent(UIMouseEvent):
 @dataclass
 class UIMouseDragEvent(UIMouseEvent):
     """Triggered when the mouse moves while one of its buttons being pressed."""
-    dx: int
-    dy: int
+    dx: float
+    dy: float
     buttons: int
     modifiers: int
